requirements.txt
playwright>=1.40.0
httpx[http2]>=0.25.0
Pillow>=10.0.0
